# Conflicts:
#	src/main/java/cc/hyperium/gui/main/HyperiumMainGui.java

+++ src/main/java/cc/hyperium/mods/motionblur/MotionBlurCommand.java
@@ -3,11 +3,7 @@ package cc.hyperium.mods.motionblur;
 import cc.hyperium.Hyperium;
 import cc.hyperium.commands.BaseCommand;
 import cc.hyperium.config.Settings;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.EntityRenderer;
-import net.minecraft.util.ChatComponentText;
 import org.apache.commons.lang3.math.NumberUtils;
 
 public class MotionBlurCommand implements BaseCommand {
@@ -37,7 +33,7 @@ public class MotionBlurCommand implements BaseCommand {
 
       if (amount != 0) {
         Settings.MOTION_BLUR_ENABLED = true;
-        Settings.MOTION_BLUR_AMOUNT = (double) amount;
+        Settings.MOTION_BLUR_AMOUNT = (float) amount;
 
         try {
           MotionBlurMod.applyShader();
